refactor: move RunResult to schema/run.ts, eliminate schema→testkit dependency inversion
