+++ .opencode/state/v0.10/prd.json
@@ -1,0 +1,13 @@
+{
+	"prdName": "v0.10-schema-testkit-inversion",
+	"description": "Fix schema↔testkit dependency inversion (Option C)",
+	"tasks": [
+		{
+			"id": "schema-testkit-inversion",
+			"title": "Move RunResult to schema, loosen ErasedCommand._execute signature",
+			"description": "Move RunResult to schema/run.ts. Loosen ErasedCommand._execute options to Readonly<Record<string, unknown>>. Update imports in command.ts, testkit/index.ts, cli/index.ts. Re-export RunResult from schema barrel and main barrel.",
+			"priority": "architecture",
+			"passes": true
+		}
+	]
+}
