+++ .opencode/state/v0.10/progress.txt
@@ -1,0 +1,25 @@
+# Progress Log
+PRD: v0.10-schema-testkit-inversion
+Started: 2026-02-11
+
+## Codebase Patterns
+- `RunResult` now canonical in `schema/run.ts` — deps only on `CLIError` (errors/) and `ActivityEvent` (schema/activity)
+- `ErasedCommand._execute` options param is `Readonly<Record<string, unknown>>` — loose at schema boundary, validated at call site
+- testkit re-exports `RunResult` from schema for public API continuity
+- Main barrel `src/index.ts` also re-exports `RunResult` (new: previously only in testkit)
+
+---
+<!-- Task logs below - APPEND ONLY -->
+
+## Task - schema-testkit-inversion
+- Created `src/core/schema/run.ts` (~47 lines) — `RunResult` interface with JSDoc
+- Updated `schema/command.ts` — import `RunResult` from `./run.js`, removed `import type` from testkit, loosened `_execute` options to `Readonly<Record<string, unknown>>`
+- Updated `schema/index.ts` — added `RunResult` re-export from `./run.js`
+- Updated `testkit/index.ts` — removed inline `RunResult` definition, added `import type { RunResult } from '../schema/run.js'`, removed unused `ActivityEvent` import
+- Updated `cli/index.ts` — split import: `RunResult` from `../schema/run.js`, `RunOptions` from `../testkit/index.js`
+- Updated `src/index.ts` — added `RunResult` to main barrel re-exports
+- Updated AGENTS.md (root, schema/, testkit/) to reflect new structure
+- Files added: `src/core/schema/run.ts`
+- Files modified: `src/core/schema/command.ts`, `src/core/schema/index.ts`, `src/core/testkit/index.ts`, `src/core/cli/index.ts`, `src/index.ts`, `AGENTS.md`, `src/core/schema/AGENTS.md`, `src/core/testkit/AGENTS.md`
+- All 1695 tests pass, tsgo clean, biome clean, dprint clean, build+publint+attw clean
+- **Learnings:** RunResult's deps (CLIError, ActivityEvent) are all at or below schema level — natural home. RunOptions can't move to schema (deps on prompt/output/help above schema). Maximally loose `Record<string, unknown>` for the _execute options param is safe because it's @internal with a single call site in cli dispatch.
